Ajout de javadoc pour la nouvelle méthode crée + Fix de bug sur LabelStar

+++ be-graphes-model/src/main/java/org/insa/graphs/model/LabelStar.java
@@ -12,7 +12,7 @@ public class LabelStar extends Label{
     /**
      * Getter de TotalCout
      */
-    public float getTotalCout() {
+    public double getTotalCout() {
         return this.getCout() + this.estimCout;
     }
